Added Functional Interface
